Refactor Employee Management System
- Removed sensitive database credentials from .env and config files.
- Cleaned up EmployeeController by removing unused methods and commented-out code.
- Updated EmployeeSalaryController to fetch salary details using DB joins.
- Added EmployeeContactController to handle employee contact details.
- Introduced EmployeeSelectController for handling employee selection via AJAX.
- Updated views to reflect changes in data structure and removed unnecessary HTML elements.
- Commented out unused routes in web.php for better clarity and organization.
- Enhanced validation and error handling in user creation process.
