+++ routes/web.php
@@ -53,18 +53,19 @@ Route::get('/getcoparatedashboard_EmployeeChart', 'CorporatedashboardController@
 Route::resource('FingerprintDevice', 'FingerprintDeviceController');
 //Route::post('addFingerprintDevice',['uses' => 'FingerprintDeviceController@store', 'as' => 'addFingerprintDevice']); 
 
+// separate employee module
 Route::get('EmloyeeList',['uses' => 'EmployeeController@employeelist', 'as' => 'EmloyeeList']); 
 Route::get('addEmployee',['uses' => 'EmployeeController@index', 'as' => 'addEmployee']);
-Route::get('employee_list_dt',['uses' => 'EmployeeController@employee_list_dt', 'as' => 'employee_list_dt']);
+// Route::get('employee_list_dt',['uses' => 'EmployeeController@employee_list_dt', 'as' => 'employee_list_dt']);
 Route::post('empoyeeUpdate',['uses' => 'EmployeeController@edit', 'as' => 'empoyeeUpdate']); 
 Route::post('empoyeeRegister',['uses' => 'EmployeeController@store', 'as' => 'empoyeeRegister']); 
-Route::post('addUserLogin',['uses' => 'EmployeeController@usercreate', 'as' => 'addUserLogin']); 
+// Route::post('addUserLogin',['uses' => 'EmployeeController@usercreate', 'as' => 'addUserLogin']); 
 Route::get('EmployeeDestroy/destroy/{id}', 'EmployeeController@destroy');
 Route::get('exportEmpoloyee', 'EmployeeController@exportempoloyee')->name('exportEmpoloyee');
 Route::get('/viewEmployee/{id}',['uses' => 'EmployeeController@show', 'as' => 'viewEmployee']);
-Route::get('/contactDetails/{id}',['uses' => 'EmployeeController@showcontact', 'as' => 'contactDetails']);
-Route::post('contactUpdate',['uses' => 'EmployeeController@editcontact', 'as' => 'contactUpdate']);
-Route::get('/viewEmergencyContacts/{id}',['uses' => 'EmployeeController@showcontact', 'as' => 'viewEmergencyContacts']);
+// Route::get('/contactDetails/{id}',['uses' => 'EmployeeController@showcontact', 'as' => 'contactDetails']);
+// Route::post('contactUpdate',['uses' => 'EmployeeController@editcontact', 'as' => 'contactUpdate']);
+// Route::get('/viewEmergencyContacts/{id}',['uses' => 'EmployeeController@showcontact', 'as' => 'viewEmergencyContacts']);
 
 Route::post('empoyeeAttachment',['uses' => 'EmployeeAttachmentController@create', 'as' => 'empoyeeAttachment']);
 Route::post('contactAttachment',['uses' => 'EmployeeAttachmentController@createcontact', 'as' => 'contactAttachment']);
@@ -524,9 +525,9 @@ Route::post('OccupationGroup.update_manual', 'OccupationGroupController@update_m
 Route::get('OccupationGroup.fetch_single', 'OccupationGroupController@fetch_single')->name('OccupationGroup.fetch_single');
 
 /*-- common routes --*/
-Route::get('employee_list_sel2', 'EmployeeController@employee_list_sel2')->name('employee_list_sel2');
-Route::get('location_list_sel2', 'EmployeeController@location_list_sel2')->name('location_list_sel2');
-Route::post('get_dept_emp_list', 'EmployeeController@get_dept_emp_list')->name('get_dept_emp_list');
+// Route::get('employee_list_sel2', 'EmployeeController@employee_list_sel2')->name('employee_list_sel2');
+// Route::get('location_list_sel2', 'EmployeeController@location_list_sel2')->name('location_list_sel2');
+// Route::post('get_dept_emp_list', 'EmployeeController@get_dept_emp_list')->name('get_dept_emp_list');
 
 
 
@@ -1389,6 +1390,19 @@ Route::post('end_user_letterprintdata', 'EndUserletterPDFController@printdata')-
 
 /*-- common employee letter routes --*/
 Route::get('employee_list_letter', 'EmployeeletterController@employee_list_letter')->name('employee_list_letter');
+
+// new employee separation routes
+Route::post('addUserLogin',['uses' => 'UserController@usercreate', 'as' => 'addUserLogin']); 
+Route::get('/contactDetails/{id}',['uses' => 'EmployeeContactController@showcontact', 'as' => 'contactDetails']);
+Route::get('/viewEmergencyContacts/{id}',['uses' => 'EmployeeContactController@showcontact', 'as' => 'viewEmergencyContacts']);
+Route::post('contactUpdate',['uses' => 'EmployeeContactController@editcontact', 'as' => 'contactUpdate']);
+
+/*-- common routes --*/
+Route::get('employee_list_sel2', 'EmployeeSelectController@employee_list_sel2')->name('employee_list_sel2');
+Route::get('location_list_sel2', 'EmployeeSelectController@location_list_sel2')->name('location_list_sel2');
+Route::post('get_dept_emp_list', 'EmployeeSelectController@get_dept_emp_list')->name('get_dept_emp_list');
+
+
 
 Route::get('/clear-cache', function() {
     Artisan::call('cache:clear');
